adding report controller

+++ Backend/koi_care_system/src/main/java/com/swpproject/koi_care_system/service/growthhistory/GrowthHistoryService.java
@@ -33,7 +33,7 @@ public class GrowthHistoryService implements IGrowthHistoryService {
         KoiFish koiFish = koiFishRepository.findById(growthCreateRequest.getKoiFishId()).orElseThrow(() -> new AppException(ErrorCode.KOI_FISH_NOT_FOUND));
         GrowthHistory growthHistory = growthHistoryMapper.mapToGrowthHistory(growthCreateRequest);
         if (growthCreateRequest.getFile() != null) {
-            growthHistory.setImageUrl(!growthCreateRequest.getFile().isEmpty() ? imageStorage.uploadImage(growthCreateRequest.getFile()) : "https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png");
+            growthHistory.setImageUrl(imageStorage.uploadImage(growthCreateRequest.getFile()));
         }
         else
             growthHistory.setImageUrl("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png");
@@ -41,10 +41,10 @@ public class GrowthHistoryService implements IGrowthHistoryService {
         GrowthHistory savedGrowthHistory = growthHistoryRepository.save(growthHistory);
         long latestId = growthHistoryRepository.findLatestByKoiFishId(koiFish.getId());
         if (savedGrowthHistory.getId() == latestId) {
-            if(savedGrowthHistory.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png"))
-                savedGrowthHistory.setImageUrl(koiFish.getImageUrl());
+            String imageUrl = (savedGrowthHistory.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png")?
+                    koiFish.getImageUrl() : savedGrowthHistory.getImageUrl());
             //Update KoiFish
-            updateKoiFish(savedGrowthHistory);
+            updateKoiFish(savedGrowthHistory,imageUrl);
             koiFishRepository.save(koiFish);
         }
         return growthHistoryMapper.mapToGrowthHistoryDto(savedGrowthHistory);
@@ -55,21 +55,19 @@ public class GrowthHistoryService implements IGrowthHistoryService {
         GrowthHistory growthHistory = growthHistoryRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.GROWTH_HISTORY_NOT_FOUND));
         growthHistoryMapper.updateGrowthHistory(growthHistory, growthUpdateRequest);
         if (growthUpdateRequest.getFile() != null)
-            if(!growthUpdateRequest.getFile().isEmpty()){
-                try {
-                    if(!growthHistory.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png"))
-                        imageStorage.deleteImage(growthHistory.getImageUrl());
-                    growthHistory.setImageUrl(imageStorage.uploadImage(growthUpdateRequest.getFile()));
-                } catch (Exception e) {
-                    throw new RuntimeException(e);
-                }
+            try {
+                if(!growthHistory.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png"))
+                    imageStorage.deleteImage(growthHistory.getImageUrl());
+                growthHistory.setImageUrl(imageStorage.uploadImage(growthUpdateRequest.getFile()));
+            } catch (Exception e) {
+                throw new RuntimeException(e);
             }
         GrowthHistory updatedGrowthHistory = growthHistoryRepository.save(growthHistory);
         KoiFish koiFish = growthHistory.getKoiFish();
         long latestId = growthHistoryRepository.findLatestByKoiFishId(koiFish.getId());
         //Update latest KoiFish
         GrowthHistory growthHistoryLatest = growthHistoryRepository.findById(latestId).orElseThrow(() -> new AppException(ErrorCode.GROWTH_HISTORY_NOT_FOUND));
-        updateKoiFish(growthHistoryLatest);
+        updateKoiFish(growthHistoryLatest,(growthHistoryLatest.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png")? koiFish.getImageUrl() : growthHistoryLatest.getImageUrl()));
         koiFishRepository.save(koiFish);
 
         return growthHistoryMapper.mapToGrowthHistoryDto(updatedGrowthHistory);
@@ -85,7 +83,8 @@ public class GrowthHistoryService implements IGrowthHistoryService {
         }
         if(!growthHistory.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png"))
             try{
-                imageStorage.deleteImage(growthHistory.getImageUrl());
+                if(!koiFish.getImageUrl().equals(growthHistory.getImageUrl()))
+                    imageStorage.deleteImage(growthHistory.getImageUrl());
             }catch (Exception e){
                 throw new RuntimeException(e);
             }
@@ -93,8 +92,7 @@ public class GrowthHistoryService implements IGrowthHistoryService {
 
         long latestId = growthHistoryRepository.findLatestByKoiFishId(koiFish.getId());
         GrowthHistory growthHistoryLatest = growthHistoryRepository.findById(latestId).orElseThrow(() -> new AppException(ErrorCode.GROWTH_HISTORY_NOT_FOUND));
-
-        updateKoiFish(growthHistoryLatest);
+        updateKoiFish(growthHistoryLatest,growthHistoryLatest.getImageUrl().equals("https://koicareimage.blob.core.windows.net/koicarestorage/defaultGrowthHistory.png")? koiFish.getImageUrl() : growthHistoryLatest.getImageUrl());
         growthHistoryRepository.save(growthHistoryLatest);
     }
 
@@ -109,9 +107,9 @@ public class GrowthHistoryService implements IGrowthHistoryService {
         return growHistories.stream().map(growthHistoryMapper::mapToGrowthHistoryDto).toList();
     }
 
-    private void updateKoiFish(GrowthHistory growthHistory) {
+    private void updateKoiFish(GrowthHistory growthHistory, String imageUrl) {
         KoiFish koiFish = growthHistory.getKoiFish();
-        koiFish.setImageUrl(growthHistory.getImageUrl());
+        koiFish.setImageUrl(imageUrl);
         koiFish.setPhysique(growthHistory.getPhysique());
         koiFish.setLength(growthHistory.getLength());
         koiFish.setWeight(growthHistory.getWeight());
